+++ imagedl/modules/sources/__init__.py
@@ -23,6 +23,7 @@ from .duckduckgo import DuckduckgoImageClient
 from .freeimages import FreeImagesImageClient
 from .everypixel import EverypixelImageClient
 from .foodiesfeed import FoodiesfeedImageClient
+from .gratisography import GratisoGraphyImageClient
 from .freenaturestock import FreeNatureStockImageClient
 
 
@@ -34,7 +35,7 @@ class ImageClientBuilder(BaseModuleBuilder):
         'I360ImageClient': I360ImageClient,              'FreeNatureStockImageClient': FreeNatureStockImageClient,    'SogouImageClient': SogouImageClient,             'YandexImageClient': YandexImageClient,
         'EverypixelImageClient': EverypixelImageClient,  'GoogleImageClient': GoogleImageClient,                      'SafebooruImageClient': SafebooruImageClient,     'YahooImageClient': YahooImageClient,
         'GelbooruImageClient': GelbooruImageClient,      'PexelsImageClient': PexelsImageClient,                      'PixabayImageClient': PixabayImageClient,         'DimTownImageClient': DimTownImageClient, 
-        'StockSnapImageClient': StockSnapImageClient,    'FreeImagesImageClient': FreeImagesImageClient,              'LifeOfPixImageClient': LifeOfPixImageClient,
+        'StockSnapImageClient': StockSnapImageClient,    'FreeImagesImageClient': FreeImagesImageClient,              'LifeOfPixImageClient': LifeOfPixImageClient,     'GratisoGraphyImageClient': GratisoGraphyImageClient,
     }